+++ CHANGELOG.md
@@ -19,6 +19,7 @@
 - On X11, if RANDR based scale factor is higher than 20 reset it to 1
 - On Wayland, add an enabled-by-default feature called `wayland-dlopen` so users can opt out of using `dlopen` to load system libraries.
 - **Breaking:** On Android, bump `ndk` and `ndk-glue` to 0.4.
+- On Windows, increase wait timer resolution for more accurate timing when using `WaitUntil`.
 
 # 0.25.0 (2021-05-15)